fix UI bookings user

+++ frontend/src/components/currentBooking/CurrentBooking.jsx
@@ -3,6 +3,7 @@ import BookedTicket from "./BookedTicket";
 import getBookedTicketsRequest from "../../clientRequest/BookedTicketsRequest.jsx";
 import "../currentBooking/CurrentBooking.css";
 import axios from "axios";
+import Swal from "sweetalert2";
 
 const CurrentBooking = () => {
   
@@ -14,10 +15,25 @@ const CurrentBooking = () => {
 
   const handleCancel = async (ticketId) => {
     try {
+      const result = await Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, cancel it!",
+      });
+  
+      if (!result.isConfirmed) {
+        return; 
+      }
+  
       console.log(ticketId);
       const response = await axios.delete(`http://localhost:5000/api/tickets/delete-ticket/${ticketId}`);
       if (response.status === 200) {
         setBookedTickets((prevTickets) => prevTickets.filter(ticket => ticket.ticketId !== ticketId));
+        // getBookedTicketsRequest(setBookedTickets)
       }
     } catch (error) {
       console.error("Failed to cancel ticket:", error);
@@ -32,7 +48,8 @@ const CurrentBooking = () => {
 
   return (
 
-    <div className="overflow-x-auto">
+    <div className="max-w-[1400px] pb-[30px] mx-auto"> 
+    <h1 className="text-[2.7em] font-bold text-white mb-[30px] mt-[20px] text-center ">Current Booking</h1>
     <table className="min-w-full bg-white rounded-lg shadow-md border border-gray-200">
       <thead className="bg-gray-200 text-gray-700 text-sm uppercase font-semibold">
         <tr>
@@ -43,6 +60,7 @@ const CurrentBooking = () => {
           <th className="py-4 px-6 text-center">Seat Class</th>
           <th className="py-4 px-6 text-center">Seat Number</th>
           <th className="py-4 px-6 text-center">Total Price</th>
+          <th className="py-4 px-6 text-center">Status</th>
           <th className="py-4 px-6 text-center">Action</th>
         </tr>
       </thead>
